bambu_ai_assistant/requirements.txt
# CREATE NEW FILE: requirements.txt
# Install these dependencies with: pip install -r requirements.txt

customtkinter>=5.2.0
opencv-python>=4.8.0
pyautogui>=0.9.54
pygetwindow>=0.0.9
Pillow>=10.0.0
pytesseract>=0.3.10
numpy>=1.24.0
requests>=2.0

# For more advanced AI vision (optional)
# ultralytics>=8.0.0  # For YOLO object detection
# torch>=2.0.0        # For deep learning models

